+++ WebSchool.Common/Constants/AssignmentConstants.cs
@@ -2,13 +2,13 @@
 {
     public static class AssignmentConstants
     {
-        public const int MinimumTitleLength = 3;
+        public const byte MinimumTitleLength = 3;
 
-        public const int MaximumTitleLength = 200;
+        public const byte MaximumTitleLength = 200;
 
-        public const int MinimumQuestionsCount = 1;
+        public const byte MinimumQuestionsCount = 1;
 
-        public const int MaximumQuestionsCount = 100;
+        public const byte MaximumQuestionsCount = 100;
 
         public const string TitleLengthMessage = "Title must be between 3 and 200 symbols";
